move method in utils module
