+++ src/screens/checkIn/checkInContainer.jsx
@@ -564,12 +564,11 @@ class CheckInContainer extends Component {
   // support
   /*-----------------------------------------------------------------------------------*/
 
-    /**
+  /**
    * shows a confirmation-dialog. if confirmed, it deletes the local data, logs the user
    * out and navigates back to the landing-screen.
    */
   deleteLocalDataAndLogout = () => {
-
     const { actions, navigation } = this.props;
     Alert.alert(
       config.text.generic.warning,
@@ -593,32 +592,6 @@ class CheckInContainer extends Component {
       ],
       { cancelable: false }
     );
-  };
-
-  /**
-   * generates a date string from Date Object (dd.mm.yyyy)
-   * @param  {Date}    inputDate the input date
-   * @param  {boolean} includeTime if true: the exported string will also contain the time
-   */
-  formatDateString = (inputDate, includeTime) => {
-    const date = new Date(inputDate);
-    const numericMonth = date.getMonth() + 1;
-
-    let monthString = numericMonth.toString();
-    let dayString = date.getDate().toString();
-    let hourString = date.getHours().toString();
-    let minutesString = date.getMinutes().toString();
-
-    if (dayString.length === 1) dayString = `0${dayString}`;
-    if (monthString.length === 1) monthString = `0${monthString}`;
-    if (hourString.length === 1) hourString = `0${hourString}`;
-    if (minutesString.length === 1) minutesString = `0${minutesString}`;
-
-    let dateString = `${dayString}.${monthString}.${date.getFullYear()}`;
-    if (includeTime)
-      dateString = `${dateString}, ${hourString}:${minutesString}`;
-
-    return dateString;
   };
 
   // rendering
@@ -665,7 +638,6 @@ class CheckInContainer extends Component {
         }
         sendReport={this.sendReport}
         updateUser={this.updateUser}
-        formatDateString={this.formatDateString}
       />
     ) : (
       // if on Survey route
